refactor: move settings to their own files
