+++ packages/enmeshed_types/lib/src/contents/request_items/form_field_settings/date_form_field_settings.dart
@@ -1,0 +1,14 @@
+part of 'form_field_settings.dart';
+
+@JsonSerializable(includeIfNull: false)
+class DateFormFieldSettings extends FormFieldSettings {
+  const DateFormFieldSettings() : super(atType: 'DateFormFieldSettings');
+
+  factory DateFormFieldSettings.fromJson(Map json) => _$DateFormFieldSettingsFromJson(Map<String, dynamic>.from(json));
+
+  @override
+  Map<String, dynamic> toJson() => _$DateFormFieldSettingsToJson(this);
+
+  @override
+  List<Object?> get props => [];
+}
